pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "docx2txt>=0.8",
    "llama-index-core>=0.12.12",
    "llama-index>=0.12.12",
    "openai>=1.60.0",
    "pypdf>=5.1.0",
    "pyyaml>=6.0.2",
    "reportlab>=4.2.5",
    "requests>=2.32.3",
    "serpapi>=0.1.5",
    "streamlit>=1.41.1",
    "trafilatura>=2.0.0",
    "llama-index-readers-web>=0.3.5",
    "llama-index-embeddings-openai>=0.3.1",
    "pandas>=2.2.3",
    "pytz>=2024.2",
    "psutil>=6.1.1",
    "twilio>=9.4.5",
    "openpyxl>=3.1.5",
]
